# packages/generated/api-client-csharp/api/openapi.yaml
openapi: 3.1.0
info:
  title: Plerion
  version: 0.1.0
servers:
- url: /
security:
- oauth2:
  - openid
- bearerAuth: []
paths:
...
  /nodes:
    delete:
      operationId: delete_nodes
      parameters:
      - description: List of Ids to delete
        explode: true
        in: query
        name: ids
        required: true
        schema:
          description: List of Ids to delete
          items:
            format: uuid
            type: string
            x-schema-name: delete_nodes_param_ids_item
            nullable: false
          type: array
        style: form
      responses:
        "200":
          content:
            application/json:
              schema: {}
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Delete Nodes
      tags:
      - Default
    get:
      operationId: get_nodes
      parameters:
      - description: Optional list of Ids to filter by
        explode: true
        in: query
        name: ids
        required: false
        schema:
          items:
            format: uuid
            type: string
            x-schema-name: get_nodes_param_ids_item
            nullable: false
          type: array
          nullable: true
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/NodeRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Get Nodes
      tags:
      - Default
    patch:
      operationId: update_nodes
      parameters:
      - explode: true
        in: query
        name: allow_missing
        required: false
        schema:
          default: false
          title: Allow Missing
          type: boolean
        style: form
      requestBody:
        content:
          application/json:
            schema:
              items:
                $ref: "#/components/schemas/NodeBatchUpdate"
              type: array
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/NodeRead"
                type: array
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Update Nodes
      tags:
      - Default
    post:
      operationId: create_node
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NodeCreate"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeRead"
          description: Successful Response
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPValidationError"
          description: Validation Error
      summary: Create Node
      tags:
      - Default
components:
  schemas:
...
    HTTPValidationError:
      example:
        detail:
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
        - msg: msg
          loc:
          - ValidationError_loc_inner
          - ValidationError_loc_inner
          type: type
      properties:
        detail:
          items:
            $ref: "#/components/schemas/ValidationError"
          type: array
      title: HTTPValidationError
...
    NodeBatchUpdate:
      example:
        rotation_x: 5.962133916683182
        rotation_w: 2.3021358869347655
        label_scale: 2.027123023002322
        link: link
        active: true
        rotation_z: 0.8008281904610115
        label_type: 7
        rotation_y: 5.637376656633329
        label: label
        label_height: 3.616076749251911
        label_width: 9.301444243932576
        link_type: 4
        position_x: 7.061401241503109
        position_y: 6.027456183070403
        layer_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        position_z: 1.4658129805029452
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        rotation_z:
          nullable: true
          type: number
        position_y:
          nullable: true
          type: number
        position_z:
          nullable: true
          type: number
        rotation_x:
          nullable: true
          type: number
        rotation_y:
          nullable: true
          type: number
        rotation_w:
          nullable: true
          type: number
        position_x:
          nullable: true
          type: number
        active:
          nullable: true
          type: boolean
        name:
          nullable: true
          type: string
        layer_id:
          format: uuid
          nullable: true
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
        label_width:
          nullable: true
          type: number
        label_height:
          nullable: true
          type: number
        label_scale:
          nullable: true
          type: number
        link_type:
          nullable: true
          type: integer
        label_type:
          nullable: true
          type: integer
        link:
          nullable: true
          type: string
        label:
          nullable: true
          type: string
      required:
      - id
      title: NodeBatchUpdate
    NodeCreate:
      example:
        rotation_x: 5.962133916683182
        rotation_w: 2.3021358869347655
        label_scale: 2.027123023002322
        link: link
        active: true
        rotation_z: 0.8008281904610115
        label_type: 7
        rotation_y: 5.637376656633329
        label: label
        label_height: 3.616076749251911
        label_width: 9.301444243932576
        link_type: 4
        position_x: 7.061401241503109
        position_y: 6.027456183070403
        layer_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        position_z: 1.4658129805029452
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          nullable: true
          type: string
        rotation_z:
          title: Rotation Z
          type: number
        position_y:
          title: Position Y
          type: number
        position_z:
          title: Position Z
          type: number
        rotation_x:
          title: Rotation X
          type: number
        rotation_y:
          title: Rotation Y
          type: number
        rotation_w:
          title: Rotation W
          type: number
        position_x:
          title: Position X
          type: number
        active:
          nullable: true
          type: boolean
        name:
          title: Name
          type: string
        layer_id:
          format: uuid
          nullable: true
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
        label_width:
          nullable: true
          type: number
        label_height:
          nullable: true
          type: number
        label_scale:
          nullable: true
          type: number
        link_type:
          nullable: true
          type: integer
        label_type:
          nullable: true
          type: integer
        link:
          nullable: true
          type: string
        label:
          nullable: true
          type: string
      required:
      - name
      - position_x
      - position_y
      - position_z
      - rotation_w
      - rotation_x
      - rotation_y
      - rotation_z
      title: NodeCreate
    NodeRead:
      example:
        rotation_x: 5.962133916683182
        rotation_w: 2.3021358869347655
        label_scale: 2.027123023002322
        link: link
        created_at: 2000-01-23T04:56:07.000+00:00
        active: true
        rotation_z: 0.8008281904610115
        label_type: 7
        rotation_y: 5.637376656633329
        label: label
        label_height: 3.616076749251911
        label_width: 9.301444243932576
        link_type: 4
        position_x: 7.061401241503109
        position_y: 6.027456183070403
        layer_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        position_z: 1.4658129805029452
        updated_at: 2000-01-23T04:56:07.000+00:00
        parent_id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        name: name
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
      properties:
        id:
          format: uuid
          title: Id
          type: string
        rotation_z:
          title: Rotation Z
          type: number
        position_y:
          title: Position Y
          type: number
        position_z:
          title: Position Z
          type: number
        rotation_x:
          title: Rotation X
          type: number
        rotation_y:
          title: Rotation Y
          type: number
        created_at:
          format: date-time
          title: Created At
          type: string
        rotation_w:
          title: Rotation W
          type: number
        updated_at:
          format: date-time
          title: Updated At
          type: string
        position_x:
          title: Position X
          type: number
        active:
          title: Active
          type: boolean
        name:
          title: Name
          type: string
        layer_id:
          format: uuid
          nullable: true
          type: string
        parent_id:
          format: uuid
          nullable: true
          type: string
        label_width:
          nullable: true
          type: number
        label_height:
          nullable: true
          type: number
        label_scale:
          nullable: true
          type: number
        link_type:
          nullable: true
          type: integer
        label_type:
          nullable: true
          type: integer
        link:
          nullable: true
          type: string
        label:
          nullable: true
          type: string
      required:
      - active
      - created_at
      - id
      - name
      - position_x
      - position_y
      - position_z
      - rotation_w
      - rotation_x
      - rotation_y
      - rotation_z
      - updated_at
      title: NodeRead
...
    ValidationError:
      example:
        msg: msg
        loc:
        - ValidationError_loc_inner
        - ValidationError_loc_inner
        type: type
      properties:
        loc:
          items:
            $ref: "#/components/schemas/ValidationError_loc_inner"
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValidationError_loc_inner:
      anyOf:
      - type: string
      - type: integer
  securitySchemes: {}